deinline nmod_mpoly IO
