+++ src/nmod_mpoly/mpolyd.c
@@ -366,46 +366,6 @@ void nmod_mpoly_convert_from_nmod_mpolyd(
     TMP_END;
 }
 
-
-void nmod_mpolyd_print(nmod_mpolyd_t poly)
-{
-
-    int first = 0;
-    slong i, j;
-    slong degb_prod;
-
-    degb_prod = WORD(1);
-    for (j = 0; j < poly->nvars; j++) {
-        degb_prod *= poly->deg_bounds[j];
-    }
-
-    first = 1;
-    for (i = 0; i < degb_prod; i++) {
-        ulong k = i;
-
-        if (poly->coeffs[i] == 0)
-            continue;
-
-        if (!first)
-            printf(" + ");
-
-        flint_printf("%wu", poly->coeffs[i]);
-
-        for (j = poly->nvars - 1; j >= 0; j--)
-        {
-            ulong m = poly->deg_bounds[j];
-            ulong e = k % m;
-            k = k / m;
-            flint_printf("*x%wd^%wd", j, e);
-        }
-        FLINT_ASSERT(k == 0);
-        first = 0;
-    }
-
-    if (first)
-        flint_printf("0");
-}
-
 slong nmod_mpolyd_length(const nmod_mpolyd_t A)
 {
     slong i, j, degb_prod;
