+++ src/nmod_mpoly/mpolyn.c
@@ -52,76 +52,6 @@ int nmod_mpolyn_is_zero(nmod_mpolyn_t A, const nmod_mpoly_ctx_t ctx)
     return A->length == 0;
 }
 
-void nmod_mpolyn_print_pretty(const nmod_mpolyn_t A,
-                                   const char ** x_in, const nmod_mpoly_ctx_t ctx)
-{
-    n_poly_struct * coeff = A->coeffs;
-    slong len = A->length;
-    ulong * exp = A->exps;
-    slong bits = A->bits;
-    slong i, j, N;
-    fmpz * exponents;
-    char ** x = (char **) x_in;
-    TMP_INIT;
-
-    if (len == 0)
-    {
-        flint_printf("0");
-        return;
-    }
-
-    N = mpoly_words_per_exp(bits, ctx->minfo);
-
-    TMP_START;
-
-    if (x == NULL)
-    {
-        x = (char **) TMP_ALLOC(ctx->minfo->nvars*sizeof(char *));
-        for (i = 0; i < ctx->minfo->nvars; i++)
-        {
-            x[i] = (char *) TMP_ALLOC(((FLINT_BITS+4)/3)*sizeof(char));
-            flint_sprintf(x[i], "x%wd", i+1);
-        }
-    }
-
-    exponents = (fmpz *) TMP_ALLOC(ctx->minfo->nvars*sizeof(fmpz));
-    for (i = 0; i < ctx->minfo->nvars; i++)
-        fmpz_init(exponents + i);
-
-    for (i = 0; i < len; i++)
-    {
-        if (i > 0)
-        {
-            printf(" + ");
-        }
-
-        printf("(");
-        n_poly_print_pretty(coeff + i, "v");
-        printf(")");
-
-        mpoly_get_monomial_ffmpz(exponents, exp + N*i, bits, ctx->minfo);
-
-        for (j = 0; j < ctx->minfo->nvars; j++)
-        {
-            int cmp = fmpz_cmp_ui(exponents + j, WORD(1));
-
-            if (cmp > 0)
-            {
-                printf("*%s^", x[j]);
-                fmpz_print(exponents + j);
-            } else if (cmp == 0)
-            {
-                printf("*%s", x[j]);
-            }
-        }
-    }
-
-    for (i = 0; i < ctx->minfo->nvars; i++)
-        fmpz_clear(exponents + i);
-
-    TMP_END;
-}
-
 void nmod_mpolyn_fit_length(nmod_mpolyn_t A, slong length, const nmod_mpoly_ctx_t ctx)
 {
     slong i;
